requirements.txt
pyodbc
fastapi
uvicorn[standard]
pydantic
azure-identity
python-dotenv
pandas
azure-storage-blob
azure-identity

